Disable heterogenous support
This error happened when launching tasks with `srun` and PMI1:

*** An error occurred in MPI_Init
--------------------------------------------------------------------------
It looks like MPI_INIT failed for some reason; your parallel process is
likely to abort.  There are many reasons that a parallel process can
fail during MPI_INIT; some of which are due to configuration or environment
problems.  This failure appears to be an internal failure; here's some
additional information (which may only be relevant to an Open MPI
developer):

  PML add procs failed
  --> Returned "Not supported" (-8) instead of "Success" (0)
--------------------------------------------------------------------------

This error is produced @ompi/mca/pml/cm/pml_cm.c#L122. Actually,
PML cm checks the CPU arch are homogeneous upon all tasks, and for some
reason, the check fails. This was false in our case, then the root cause
is probably located at the initialization of this data structure.

When heterogeneous support is enable, this data structure is
initialiazed @ompi/proc/proc.c#L168. When PMI is used, this macro
retrieves the `pmix.arch` key out-of the PMI hash managed by the RM
(eg. slurm). This key is normally published @ompi/proc/proc.c#L281.
However, it sinces the set does not happen according to slurm PMI
library, ex:

  $ PMI_DEBUG=1 OMPI_MCA_pml=cm srun mpi_hello_world
  In: PMI_Init
  In: PMI_Init
  In: PMI_KVS_Get_value_length_max
  In: PMI_KVS_Get_name_length_max
  In: PMI_KVS_Get_key_length_max
  In: PMI_Get_kvs_domain_id
  In: PMI_Get_rank
  In: PMI_KVS_Get_my_name
  In: PMI_Get_clique_size
  In: PMI_Get_clique_ranks
  In: PMI_Get_universe_size
  In: PMI_Get_size
  In: PMI_Get_appnum
  In: PMI_KVS_Get_value_length_max
  In: PMI_KVS_Get_name_length_max
  In: PMI_KVS_Get_key_length_max
  In: PMI_Get_kvs_domain_id
  In: PMI_Get_rank
  In: PMI_KVS_Get_my_name
  In: PMI_KVS_Get(29687808-1-key0)
  In: PMI_Get_clique_size
  In: PMI_Get_clique_ranks
  In: PMI_Get_universe_size
  In: PMI_Get_size
  In: PMI_Get_appnum
  In: PMI_KVS_Get(29687808-0-key0)
  In: PMI_KVS_Get(29687808-0-key0)
  In: PMI_KVS_Get(29687808-1-key0)
  In: PMI_KVS_Put(29687808-0-key0:[…])
  In: PMI_KVS_Put(29687808-1-key0:[…])
  In: PMI_KVS_Commit
  In: PMI_KVS_Commit
  In: PMI_Barrier
  In: PMI_Barrier
  Past PMI_Barrier
  Past PMI_Barrier
  In: PMI_KVS_Get(29687808-0-key0)
  In: PMI_KVS_Get(29687808-1-key0)
  In: PMI_KVS_Get(29687808-0-key0)
  In: PMI_KVS_Get(29687808-1-key0)
  --------------------------------------------------------------------------
  It looks like MPI_INIT failed for some reason; your parallel process is
  likely to abort.  There are many reasons that a parallel process can
  fail during MPI_INIT; some of which are due to configuration or environment
  problems.  This failure appears to be an internal failure; here's some
  additional information (which may only be relevant to an Open MPI
  developer):

    PML add procs failed
    --> Returned "Not supported" (-8) instead of "Success" (0)
  --------------------------------------------------------------------------
  *** An error occurred in MPI_Init
  --------------------------------------------------------------------------
  It looks like MPI_INIT failed for some reason; your parallel process is
  likely to abort.  There are many reasons that a parallel process can
  fail during MPI_INIT; some of which are due to configuration or environment
  problems.  This failure appears to be an internal failure; here's some
  additional information (which may only be relevant to an Open MPI
  developer):

    PML add procs failed
    --> Returned "Not supported" (-8) instead of "Success" (0)
  --------------------------------------------------------------------------
  *** An error occurred in MPI_Init
  *** on a NULL communicator
  *** MPI_ERRORS_ARE_FATAL (processes in this communicator will now abort,
  ***    and potentially your MPI job)
  *** on a NULL communicator
  *** MPI_ERRORS_ARE_FATAL (processes in this communicator will now abort,
  ***    and potentially your MPI job)
  In: PMI_Abort(1, N/A)
  In: PMI_Abort(1, N/A)

I guess something was broken with heterogeneous support in combination
with PMI1, following adoption of PMIX into OpenMPI codebase. Since, HPC
clusters are mostly homogeneous in terms of CPU architecture, disable
heterogeneous support as a workaround for this bug.